src/training/train_darts.py
"""
Training script for DARTS (Differentiable Architecture Search)
"""
import argparse
import os
import logging
import numpy as np
import pandas as pd
from sklearn.model_selection import train_test_split
import tensorflow as tf
from tensorflow import keras
from tensorflow.keras import layers

logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)


def build_darts_model(input_shape, num_classes, architecture_str=None):
    """Build DARTS model"""
    model = keras.Sequential()
    model.add(layers.Input(shape=input_shape))
    
    # DARTS cell structure
    if architecture_str:
        # Parse DARTS architecture
        # Simplified implementation
        model.add(layers.SeparableConv2D(32, (3, 3), activation="relu", padding="same"))
        model.add(layers.SeparableConv2D(64, (3, 3), activation="relu", padding="same"))
        model.add(layers.MaxPooling2D((2, 2)))
        model.add(layers.SeparableConv2D(128, (3, 3), activation="relu", padding="same"))
        model.add(layers.GlobalAveragePooling2D())
    else:
        # Default DARTS-like architecture
        model.add(layers.SeparableConv2D(32, (3, 3), activation="relu", input_shape=input_shape))
        model.add(layers.SeparableConv2D(64, (3, 3), activation="relu"))
        model.add(layers.MaxPooling2D((2, 2)))
        model.add(layers.SeparableConv2D(128, (3, 3), activation="relu"))
        model.add(layers.GlobalAveragePooling2D())
    
    model.add(layers.Dense(num_classes if num_classes > 2 else 1,
                          activation="softmax" if num_classes > 2 else "sigmoid"))
    
    loss = "sparse_categorical_crossentropy" if num_classes > 2 else "binary_crossentropy"
    model.compile(optimizer="adam", loss=loss, metrics=["accuracy"])
    
    return model


def load_data(data_path):
    """Load data"""
    if not os.path.exists(data_path):
        X = np.random.rand(1000, 32, 32, 3)
        y = np.random.randint(0, 10, 1000)
        return X, y
    
    df = pd.read_csv(data_path)
    X = df.iloc[:, :-1].values
    y = df.iloc[:, -1].values
    
    if len(X.shape) == 2 and X.shape[1] == 3072:
        X = X.reshape(-1, 32, 32, 3)
    
    return X, y


def train_model(data_path, architecture_str=None):
    """Train DARTS model"""
    X, y = load_data(data_path)
    X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)
    
    X_train = X_train.astype("float32") / 255.0
    X_test = X_test.astype("float32") / 255.0
    
    num_classes = len(np.unique(y))
    input_shape = X_train.shape[1:]
    
    model = build_darts_model(input_shape, num_classes, architecture_str)
    
    history = model.fit(
        X_train, y_train,
        batch_size=32,
        epochs=10,
        validation_split=0.2,
        verbose=1,
    )
    
    test_loss, test_accuracy = model.evaluate(X_test, y_test, verbose=0)
    
    print(f"accuracy={test_accuracy:.4f}")
    print(f"loss={test_loss:.4f}")
    
    model_path = os.getenv("MODEL_PATH", "/tmp/model")
    os.makedirs(model_path, exist_ok=True)
    model.save(os.path.join(model_path, "model.h5"))
    
    return test_accuracy


def main():
    parser = argparse.ArgumentParser()
    parser.add_argument("--data-path", type=str, default="/data/train.csv")
    parser.add_argument("--architecture", type=str, default=None)
    
    args = parser.parse_args()
    
    train_model(
        data_path=args.data_path,
        architecture_str=args.architecture,
    )


if __name__ == "__main__":
    main()



