requirements.txt
# Core ML Libraries
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3
scipy==1.11.4

# Deep Learning
tensorflow==2.13.0
torch==2.1.0
keras==2.13.1

# Kubeflow & Katib
kubeflow-katib==0.15.0
kubeflow-pipelines==2.0.5
kfp==2.5.0
kubernetes==28.1.0

# Hyperparameter Optimization
optuna==3.4.0
hyperopt==0.2.7
scikit-optimize==0.9.0

# Model Serving
kserve==0.11.0
kfserving==0.7.0

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
mlflow==2.8.1
wandb==0.16.0

# Data Processing
joblib==1.3.2
dill==0.3.7

# API & Web
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Jupyter
jupyter==1.0.0
notebook==7.0.6
ipykernel==6.26.0



